Update Graph View while activities are being dragged

+++ scenario/src/main/java/gov/nasa/arc/mct/scenario/view/AbstractTimelineView.java
@@ -59,7 +59,7 @@ public abstract class AbstractTimelineView extends View {
 	
 	public double getPixelScale() {
 		return timelineContainer != null ?
-				(getWidth() - getLeftPadding() - getRightPadding()) / 
+				(double) (getWidth() - getLeftPadding() - getRightPadding()) / 
 				(double) (timelineContainer.getEnd() - timelineContainer.getStart()) : 
 				1.0;
 	}
